refactor(p1): extract ContactsAccessTools.kt — contacts_search/write factories
Splits the contacts_search + contacts_write tools and their two ContentResolver
helpers out of SystemAccessTools. Same factory shape as the LocalTools session:
top-level `internal fun createContactsSearchTool(context, deps): Tool` and
`internal fun createContactsWriteTool(context, deps): Tool`, plus file-private
`queryContacts(context, query, limit)` and `createContact(context, name, phone, email)`.

SystemAccessTools coordinator replaces the two `private val xTool = Tool(...)`
fields with `by lazy { createXxxTool(context, deps) }` delegators (matching the
LocalTools pattern).

Side cleanup: drop now-unused imports `android.content.ContentProviderOperation`
and `android.provider.ContactsContract` from SystemAccessTools.kt. `ContentUris`
stays — still used by calendar_create + media_search.

SystemAccessTools.kt: 1338 → 1194 lines (-144).
ContactsAccessTools.kt: 158 new.

Verification:
- :app:compileDebugKotlin --offline → BUILD SUCCESSFUL
